internal/solutions: Refactor day01 to implement Solver interface
