+++ internal/solutions/day01/spec.md
