+++ internal/solutions/day02/spec.md
